write individual components
