+++ frontend/src/components/listeners/ExtractionJobs.tsx
@@ -4,28 +4,18 @@ import Table from "@mui/material/Table";
 import TableBody from "@mui/material/TableBody";
 import TableCell from "@mui/material/TableCell";
 import TableContainer from "@mui/material/TableContainer";
-import TableHead from "@mui/material/TableHead";
 import TablePagination from "@mui/material/TablePagination";
 import TableRow from "@mui/material/TableRow";
-import TableSortLabel from "@mui/material/TableSortLabel";
-import Toolbar from "@mui/material/Toolbar";
 import Paper from "@mui/material/Paper";
-import {visuallyHidden} from "@mui/utils";
 import CancelIcon from "@mui/icons-material/Cancel";
 import CheckCircleIcon from "@mui/icons-material/CheckCircle";
 import AccessTimeIcon from "@mui/icons-material/AccessTime";
 import PlayCircleOutlineIcon from "@mui/icons-material/PlayCircleOutline";
 import {theme} from "../../theme";
-import {FormControl, InputLabel, MenuItem, Select, Typography} from "@mui/material";
-import {useState} from "react";
-import {DateRange} from "@appbaseio/reactivesearch";
-import {AdapterDayjs} from "@mui/x-date-pickers/AdapterDayjs";
-import {DateTimePicker} from "@mui/x-date-pickers/DateTimePicker";
-import {ClowderMetadataTextField} from "../styledComponents/ClowderMetadataTextField";
-import {LocalizationProvider} from "@mui/x-date-pickers/LocalizationProvider";
-import {ClowderFootnote} from "../styledComponents/ClowderFootnote";
+import {ExtractionJobsToolbar} from "./ExtractionJobsToolbar";
+import {EnhancedTableHead} from "./ExtractionJobsTableHeader";
 
-interface Data {
+export interface Data {
 	status: string;
 	jobId: string;
 	created: string;
@@ -45,7 +35,7 @@ function descendingComparator<T>(a: T, b: T, orderBy: keyof T) {
 	return 0;
 }
 
-type Order = "asc" | "desc";
+export type Order = "asc" | "desc";
 
 function getComparator<Key extends keyof any>(
 	order: Order,
@@ -72,92 +62,6 @@ function stableSort<T>(array: readonly T[], comparator: (a: T, b: T) => number) 
 	});
 	return stabilizedThis.map((el) => el[0]);
 }
-
-interface EnhancedTableProps {
-	onRequestSort: (event: React.MouseEvent<unknown>, property: keyof Data) => void;
-	order: Order;
-	orderBy: string;
-	headCells: any;
-}
-
-function EnhancedTableHead(props: EnhancedTableProps) {
-	const {order, orderBy, onRequestSort, headCells} = props;
-	const createSortHandler =
-		(property: keyof Data) => (event: React.MouseEvent<unknown>) => {
-			onRequestSort(event, property);
-		};
-
-	return (
-		<TableHead>
-			<TableRow>
-				{headCells.map((headCell) => (
-					<TableCell
-						key={headCell.id}
-						align="left"
-						padding="normal"
-						sortDirection={orderBy === headCell.id ? order : false}
-					>
-						<TableSortLabel
-							active={orderBy === headCell.id}
-							direction={orderBy === headCell.id ? order : "asc"}
-							onClick={createSortHandler(headCell.id)}
-						>
-							{headCell.label}
-							{orderBy === headCell.id ? (
-								<Box component="span" sx={visuallyHidden}>
-									{order === "desc" ? "sorted descending" : "sorted ascending"}
-								</Box>
-							) : null}
-						</TableSortLabel>
-					</TableCell>
-				))}
-			</TableRow>
-		</TableHead>
-	);
-}
-
-interface EnhancedTableToolbarProps {
-	numExecution:number;
-}
-
-const EnhancedTableToolbar = (props: EnhancedTableToolbarProps) => {
-	const {numExecution} = props;
-	const [selectedStatus, setSelectedStatus] = useState("all");
-	const [selectedCreatedTime, setSelectedCreatedTime] = useState();
-
-	return (
-		 <Box sx={{ flexGrow: 1, padding: "1em 0"}}>
-			<Toolbar>
-				<Typography sx={{ flexGrow: 1 }}>{numExecution} extractions run</Typography>
-				{/*filter by status*/}
-				<FormControl variant="standard" sx={{ m: 1, minWidth: 120 }}>
-					<InputLabel id="demo-simple-select-label">Status</InputLabel>
-					  <Select
-						labelId="demo-simple-select-label"
-						id="demo-simple-select"
-						value={selectedStatus}
-						label="Status"
-						onChange={(e: React.ChangeEvent<HTMLInputElement>)=>{setSelectedStatus(e.target.value);}}
-					  >
-						<MenuItem value="all">All</MenuItem>
-						<MenuItem value="StatusMessage.start">Start</MenuItem>
-						<MenuItem value="StatusMessage.processing">Processing</MenuItem>
-						<MenuItem value="StatusMessage.done">Done</MenuItem>
-						<MenuItem value="StatusMessage.failed">Failed</MenuItem>
-					  </Select>
-				</FormControl>
-				<LocalizationProvider dateAdapter={AdapterDayjs}>
-					<DateTimePicker
-						label="Submitted at"
-						value={selectedCreatedTime}
-						onChange={(value)=>{setSelectedCreatedTime(value);}}
-						renderInput={(props) => <ClowderMetadataTextField {...props} variant="standard"/>}
-					/>
-				</LocalizationProvider>
-			</Toolbar>
-		 </Box>
-	);
-};
 
 export const ExtractionJobs = (props) => {
 	const {rows, headCells} = props;
@@ -192,7 +96,7 @@ export const ExtractionJobs = (props) => {
 		<Box sx={{width: "100%"}}>
 			<Paper sx={{width: "100%", mb: 2}}>
 				<TableContainer>
-					<EnhancedTableToolbar numExecution={rows.length}/>
+					<ExtractionJobsToolbar numExecution={rows.length}/>
 					<Table
 						sx={{minWidth: 750}}
 						aria-labelledby="tableTitle"
@@ -207,32 +111,36 @@ export const ExtractionJobs = (props) => {
 						<TableBody>
 							{
 								stableSort(rows, getComparator(order, orderBy))
-								.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
-								.map((row) => {
-									return (
-										<TableRow key={row.jobId}>
-											<TableCell align="right" sx={{color:theme.palette.primary.main}}>
-											{
-												row.status.includes("StatusMessage.start")? <PlayCircleOutlineIcon />:null
-											}
-											{
-												row.status.includes("StatusMessage.processing") ?<AccessTimeIcon />: null
-											}
-											{
-												row.status.includes("StatusMessage.done") ? <CheckCircleIcon />: null
-											}
-											{
-												row.status.includes("StatusMessage.failed") ? <CancelIcon />: null
-											}
-											</TableCell>
-											<TableCell align="left">{row.jobId}</TableCell>
-											<TableCell align="left">{row.created}</TableCell>
-											<TableCell align="left">{row.creator}</TableCell>
-											<TableCell align="left">{row.duration}</TableCell>
-											<TableCell align="left">{row.fileId}</TableCell>
-										</TableRow>
-									);
-								})
+									.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
+									.map((row) => {
+										return (
+											<TableRow key={row.jobId}>
+												<TableCell align="right" sx={{color: theme.palette.primary.main}}>
+													{
+														row.status.includes("StatusMessage.start") ?
+															<PlayCircleOutlineIcon/> : null
+													}
+													{
+														row.status.includes("StatusMessage.processing") ?
+															<AccessTimeIcon/> : null
+													}
+													{
+														row.status.includes("StatusMessage.done") ?
+															<CheckCircleIcon/> : null
+													}
+													{
+														row.status.includes("StatusMessage.failed") ?
+															<CancelIcon/> : null
+													}
+												</TableCell>
+												<TableCell align="left">{row.jobId}</TableCell>
+												<TableCell align="left">{row.created}</TableCell>
+												<TableCell align="left">{row.creator}</TableCell>
+												<TableCell align="left">{row.duration}</TableCell>
+												<TableCell align="left">{row.fileId}</TableCell>
+											</TableRow>
+										);
+									})
 							}
 							{emptyRows > 0 && (
 								<TableRow
