+++ frontend/src/components/listeners/ExtractionJobsTableHeader.tsx
@@ -1,0 +1,51 @@
+import * as React from "react";
+import TableHead from "@mui/material/TableHead";
+import TableRow from "@mui/material/TableRow";
+import TableCell from "@mui/material/TableCell";
+import TableSortLabel from "@mui/material/TableSortLabel";
+import Box from "@mui/material/Box";
+import {visuallyHidden} from "@mui/utils";
+import {Data, Order} from "./ExtractionJobs";
+
+interface EnhancedTableProps {
+	onRequestSort: (event: React.MouseEvent<unknown>, property: keyof Data) => void;
+	order: Order;
+	orderBy: string;
+	headCells: any;
+}
+
+export const EnhancedTableHead = (props: EnhancedTableProps) => {
+	const {order, orderBy, onRequestSort, headCells} = props;
+	const createSortHandler =
+		(property: keyof Data) => (event: React.MouseEvent<unknown>) => {
+			onRequestSort(event, property);
+		};
+
+	return (
+		<TableHead>
+			<TableRow>
+				{headCells.map((headCell) => (
+					<TableCell
+						key={headCell.id}
+						align="left"
+						padding="normal"
+						sortDirection={orderBy === headCell.id ? order : false}
+					>
+						<TableSortLabel
+							active={orderBy === headCell.id}
+							direction={orderBy === headCell.id ? order : "asc"}
+							onClick={createSortHandler(headCell.id)}
+						>
+							{headCell.label}
+							{orderBy === headCell.id ? (
+								<Box component="span" sx={visuallyHidden}>
+									{order === "desc" ? "sorted descending" : "sorted ascending"}
+								</Box>
+							) : null}
+						</TableSortLabel>
+					</TableCell>
+				))}
+			</TableRow>
+		</TableHead>
+	);
+}
